Rename camel_case -> string
