+++ lib/src/generators/colors_generator.dart
@@ -5,7 +5,7 @@ import 'package:dartx/dartx.dart';
 import 'package:flutter_gen/src/generators/generator_helper.dart';
 import 'package:flutter_gen/src/settings/color_path.dart';
 import 'package:flutter_gen/src/settings/flutter_gen.dart';
-import 'package:flutter_gen/src/utils/camel_case.dart';
+import 'package:flutter_gen/src/utils/string.dart';
 import 'package:flutter_gen/src/utils/color.dart';
 import 'package:path/path.dart';
 import 'package:xml/xml.dart';
