+++ lib/src/utils/string.dart
